Use min and max of the C++ STL
Based on TortoiseGit commit 59ead584bcde1637166b1c87be8b54ec37fb3a69.

git-svn-id: http://svn.osdn.net/svnroot/tortoisesvn/trunk@28435 154be670-7969-47af-9ab0-c6fb378d2295

+++ src/SVN/SVNStatusListCtrl.cpp
@@ -6546,7 +6546,7 @@ LRESULT CSVNStatusListCtrl::OnRefreshStatusMsg(WPARAM, LPARAM)
 bool CSVNStatusListCtrl::CheckMultipleDiffs()
 {
     UINT selCount = GetSelectedCount();
-    if (selCount > max(3, (DWORD)CRegDWORD(L"Software\\TortoiseSVN\\NumDiffWarning", 10)))
+    if (selCount > max((DWORD)3, (DWORD)CRegDWORD(L"Software\\TortoiseSVN\\NumDiffWarning", 10)))
     {
         CString message;
         message.Format(CString(MAKEINTRESOURCE(IDS_STATUSLIST_WARN_MAXDIFF)), selCount);
