+++ src/TortoiseProc/RevisionGraph/RevisionGraphDlgFunc.cpp
@@ -81,8 +81,8 @@ void CRevisionGraphWnd::BuildPreview()
 
     // make sure the preview window has a minimal size
 
-    m_previewWidth = (int)min (max (graphRect.Width() * m_previewZoom, 30), preViewSize.cx);
-    m_previewHeight = (int)min (max (graphRect.Height() * m_previewZoom, 30), preViewSize.cy);
+    m_previewWidth  = min((int)max(graphRect.Width() * m_previewZoom, 30.0f), (int)preViewSize.cx);
+    m_previewHeight = min((int)max(graphRect.Height() * m_previewZoom, 30.0f), (int)preViewSize.cy);
 
     CClientDC ddc(this);
     CDC dc;
@@ -402,7 +402,7 @@ void CRevisionGraphWnd::DoZoom (float fZoomFactor, bool updateScrollbars)
 
     m_nFontSize = max(1, int(DEFAULT_ZOOM_FONT * fZoomFactor));
     if (m_nFontSize < SMALL_ZOOM_FONT_THRESHOLD)
-        m_nFontSize = min (SMALL_ZOOM_FONT_THRESHOLD, int(SMALL_ZOOM_FONT * fZoomFactor));
+        m_nFontSize = min((int)SMALL_ZOOM_FONT_THRESHOLD, int(SMALL_ZOOM_FONT * fZoomFactor));
 
     for (int i=0; i<MAXFONTS; i++)
     {
